issue 494
