+++ src/sempy_labs/report/_report_functions.py
@@ -2,7 +2,6 @@ import sempy.fabric as fabric
 import pandas as pd
 import json
 import os
-import time
 import copy
 from anytree import Node, RenderTree
 from powerbiclient import Report
@@ -10,7 +9,6 @@ from pyspark.sql.functions import col, flatten
 from sempy_labs.report._generate_report import update_report_from_reportjson
 from sempy_labs.lakehouse._lakehouse import lakehouse_attached
 from sempy_labs._helper_functions import (
-    generate_embedded_filter,
     resolve_report_id,
     resolve_lakehouse_name,
     language_validate,
@@ -143,254 +141,6 @@ def report_dependency_tree(workspace: Optional[str | UUID] = None):
     # Print the tree structure
     for pre, _, node in RenderTree(node_dict[workspace_name]):
         print(f"{pre}{node.custom_property}'{node.name}'")
-
-
-@log
-def export_report(
-    report: str,
-    export_format: str,
-    file_name: Optional[str] = None,
-    bookmark_name: Optional[str] = None,
-    page_name: Optional[str] = None,
-    visual_name: Optional[str] = None,
-    report_filter: Optional[str] = None,
-    workspace: Optional[str | UUID] = None,
-):
-    """
-    Exports a Power BI report to a file in your lakehouse.
-
-    This is a wrapper function for the following APIs: `Reports - Export To File In Group <https://learn.microsoft.com/rest/api/power-bi/reports/export-to-file-in-group>`_, `Reports - Get Export To File Status In Group <https://learn.microsoft.com/rest/api/power-bi/reports/get-export-to-file-status-in-group>`_, `Reports - Get File Of Export To File In Group <https://learn.microsoft.com/rest/api/power-bi/reports/get-file-of-export-to-file-in-group>`_.
-
-    Parameters
-    ----------
-    report : str
-        Name of the Power BI report.
-    export_format : str
-        The format in which to export the report. For image formats, enter the file extension in this parameter, not 'IMAGE'.
-        `Valid formats <https://learn.microsoft.com/rest/api/power-bi/reports/export-to-file-in-group#fileformat>`_
-    file_name : str, default=None
-        The name of the file to be saved within the lakehouse. Do not include the file extension. Defaults ot the reportName parameter value.
-    bookmark_name : str, default=None
-        The name (GUID) of a bookmark within the report.
-    page_name : str, default=None
-        The name (GUID) of the report page.
-    visual_name : str, default=None
-        The name (GUID) of a visual. If you specify this parameter you must also specify the page_name parameter.
-    report_filter : str, default=None
-        A report filter to be applied when exporting the report. Syntax is user-friendly. See above for examples.
-    workspace : str | uuid.UUID, default=None
-        The Fabric workspace name or ID.
-        Defaults to None which resolves to the workspace of the attached lakehouse
-        or if no lakehouse attached, resolves to the workspace of the notebook.
-    """
-
-    if not lakehouse_attached():
-        raise ValueError(
-            f"{icons.red_dot} In order to run the 'export_report' function, a lakehouse must be attached to the notebook. Please attach a lakehouse to this notebook."
-        )
-
-    (workspace_name, workspace_id) = resolve_workspace_name_and_id(workspace)
-
-    if isinstance(page_name, str):
-        page_name = [page_name]
-    if isinstance(visual_name, str):
-        visual_name = [visual_name]
-
-    if bookmark_name is not None and (page_name is not None or visual_name is not None):
-        raise ValueError(
-            f"{icons.red_dot} If the 'bookmark_name' parameter is set, the 'page_name' and 'visual_name' parameters must not be set."
-        )
-
-    if visual_name is not None and page_name is None:
-        raise ValueError(
-            f"{icons.red_dot} If the 'visual_name' parameter is set, the 'page_name' parameter must be set."
-        )
-
-    validFormats = {
-        "ACCESSIBLEPDF": ".pdf",
-        "CSV": ".csv",
-        "DOCX": ".docx",
-        "MHTML": ".mhtml",
-        "PDF": ".pdf",
-        "PNG": ".png",
-        "PPTX": ".pptx",
-        "XLSX": ".xlsx",
-        "XML": ".xml",
-        "BMP": ".bmp",
-        "EMF": ".emf",
-        "GIF": ".gif",
-        "JPEG": ".jpeg",
-        "TIFF": ".tiff",
-    }
-
-    export_format = export_format.upper()
-    fileExt = validFormats.get(export_format)
-    if fileExt is None:
-        raise ValueError(
-            f"{icons.red_dot} The '{export_format}' format is not a valid format for exporting Power BI reports. Please enter a valid format. Options: {validFormats}"
-        )
-
-    if file_name is None:
-        file_name = f"{report}{fileExt}"
-    else:
-        file_name = f"{file_name}{fileExt}"
-
-    folderPath = "/lakehouse/default/Files"
-    filePath = os.path.join(folderPath, file_name)
-
-    dfI = fabric.list_items(workspace=workspace_id)
-    dfI_filt = dfI[
-        (dfI["Type"].isin(["Report", "PaginatedReport"]))
-        & (dfI["Display Name"] == report)
-    ]
-
-    if len(dfI_filt) == 0:
-        raise ValueError(
-            f"{icons.red_dot} The '{report}' report does not exist in the '{workspace_name}' workspace."
-        )
-
-    reportType = dfI_filt["Type"].iloc[0]
-
-    # Limitations
-    pbiOnly = ["PNG"]
-    paginatedOnly = [
-        "ACCESSIBLEPDF",
-        "CSV",
-        "DOCX",
-        "BMP",
-        "EMF",
-        "GIF",
-        "JPEG",
-        "TIFF",
-        "MHTML",
-        "XLSX",
-        "XML",
-    ]
-
-    if reportType == "Report" and export_format in paginatedOnly:
-        raise ValueError(
-            f"{icons.red_dot} The '{export_format}' format is only supported for paginated reports."
-        )
-
-    if reportType == "PaginatedReport" and export_format in pbiOnly:
-        raise ValueError(
-            f"{icons.red_dot} The '{export_format}' format is only supported for Power BI reports."
-        )
-
-    if reportType == "PaginatedReport" and (
-        bookmark_name is not None or page_name is not None or visual_name is not None
-    ):
-        raise ValueError(
-            f"{icons.red_dot} Export for paginated reports does not support bookmarks/pages/visuals. Those parameters must not be set for paginated reports."
-        )
-
-    reportId = dfI_filt["Id"].iloc[0]
-
-    if (
-        export_format in ["BMP", "EMF", "GIF", "JPEG", "TIFF"]
-        and reportType == "PaginatedReport"
-    ):
-        request_body = {
-            "format": "IMAGE",
-            "paginatedReportConfiguration": {
-                "formatSettings": {"OutputFormat": export_format.lower()}
-            },
-        }
-    elif bookmark_name is None and page_name is None and visual_name is None:
-        request_body = {"format": export_format}
-    elif bookmark_name is not None:
-        if reportType == "Report":
-            request_body = {
-                "format": export_format,
-                "powerBIReportConfiguration": {
-                    "defaultBookmark": {"name": bookmark_name}
-                },
-            }
-    elif page_name is not None and visual_name is None:
-        if reportType == "Report":
-            request_body = {"format": export_format, "powerBIReportConfiguration": {}}
-
-            request_body["powerBIReportConfiguration"]["pages"] = []
-            dfPage = list_report_pages(report=report, workspace=workspace_id)
-
-            for page in page_name:
-                dfPage_filt = dfPage[dfPage["Page ID"] == page]
-                if len(dfPage_filt) == 0:
-                    raise ValueError(
-                        f"{icons.red_dot} The '{page}' page does not exist in the '{report}' report within the '{workspace_name}' workspace."
-                    )
-
-                page_dict = {"pageName": page}
-                request_body["powerBIReportConfiguration"]["pages"].append(page_dict)
-
-    elif page_name is not None and visual_name is not None:
-        if len(page_name) != len(visual_name):
-            raise ValueError(
-                f"{icons.red_dot} Each 'visual_name' must map to a single 'page_name'."
-            )
-
-        if reportType == "Report":
-            request_body = {"format": export_format, "powerBIReportConfiguration": {}}
-
-            request_body["powerBIReportConfiguration"]["pages"] = []
-            dfVisual = list_report_visuals(report=report, workspace=workspace_id)
-            a = 0
-            for page in page_name:
-                visual = visual_name[a]
-
-                dfVisual_filt = dfVisual[
-                    (dfVisual["Page ID"] == page) & (dfVisual["Visual ID"] == visual)
-                ]
-                if len(dfVisual_filt) == 0:
-                    raise ValueError(
-                        f"{icons.red_dot} The '{visual}' visual does not exist on the '{page}' in the '{report}' report within the '{workspace_name}' workspace."
-                    )
-
-                page_dict = {"pageName": page, "visualName": visual}
-                request_body["powerBIReportConfiguration"]["pages"].append(page_dict)
-                a += 1
-
-    # Transform and add report filter if it is specified
-    if report_filter is not None and reportType == "Report":
-        reportFilter = generate_embedded_filter(filter=report_filter)
-        report_level_filter = {"filter": reportFilter}
-
-        if "powerBIReportConfiguration" not in request_body:
-            request_body["powerBIReportConfiguration"] = {}
-        request_body["powerBIReportConfiguration"]["reportLevelFilters"] = [
-            report_level_filter
-        ]
-
-    base_url = f"/v1.0/myorg/groups/{workspace_id}/reports/{reportId}"
-    response = _base_api(
-        request=f"{base_url}/ExportTo",
-        method="post",
-        payload=request_body,
-        status_codes=202,
-    )
-    export_id = json.loads(response.content).get("id")
-
-    get_status_url = f"{base_url}/exports/{export_id}"
-    response = _base_api(request=get_status_url, status_codes=[200, 202])
-    response_body = json.loads(response.content)
-    while response_body["status"] not in ["Succeeded", "Failed"]:
-        time.sleep(3)
-        response = _base_api(request=get_status_url, status_codes=[200, 202])
-        response_body = json.loads(response.content)
-    if response_body["status"] == "Failed":
-        raise ValueError(
-            f"{icons.red_dot} The export for the '{report}' report within the '{workspace_name}' workspace in the '{export_format}' format has failed."
-        )
-    else:
-        response = _base_api(request=f"{get_status_url}/file")
-        print(
-            f"{icons.in_progress} Saving the '{export_format}' export for the '{report}' report within the '{workspace_name}' workspace to the lakehouse..."
-        )
-        with open(filePath, "wb") as export_file:
-            export_file.write(response.content)
-        print(
-            f"{icons.green_dot} The '{export_format}' export for the '{report}' report within the '{workspace_name}' workspace has been saved to the following location: '{filePath}'."
-        )
 
 
 def clone_report(
